some things shouldn't be told

+++ lab6/Graph/Point.cs
@@ -1,3 +1,3 @@
 namespace lab6.Graph;
 
-public record Point(float X, float Y);
+public record Point(double X, double Y);
